backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.51.0
pydantic==2.5.2
tiktoken==0.7.0 